delete method added

+++ 04_serlvets/servlets/src/main/java/ru/netology/service/PostService.java
@@ -27,7 +27,7 @@ public class PostService {
     return repository.save(post);
   }
 
-  public Post edit(int id, Post post) { return repository.edit(id, post); }
+  public Post edit(long id, Post post) { return repository.edit(id, post); }
 
   public void removeById(long id) {
     repository.removeById(id);
